Sign

+++ 0x02-functions_nested_loops/5-sign.c
@@ -14,11 +14,11 @@
 
 int print_sign(int n)
 {
-	if (n > 0)
+	if ((int)n > 0)
 	{
 		return (1);
 	}
-	else if (n == 0)
+	else if (int(n) == 0)
 	{
 		return (0);
 	}
